Backport 6516a35 "fix resolution test" from time
The upstream version of time was incorrect in newer versions until
https://github.com/haskell/time/commit/6516a35b3d8cad14a21a559a5d2fc22ddda32fcc.
Because the compat implementation was correct and the upstream version
of time is now logically identical, simply replace both versions with
the current upstream time implementation.

See https://github.com/haskellari/time-compat/issues/31 for details on
which platforms the distinction matters.  Does NOT close that issue
though, because the original issue is for platforms on which the actual
resolution is less precise than the expected, and this will only fix
platforms on which the actual resolution is more precise.